refactor: 클래스 파일 옮기기 수행
